refactor: move BlackConfig and filter_python_files to more logical places
